Update function.php, home.php and upload_item.php
functions.php:
change the create_new_user_in_filesystem with new folder names.
modify is_allowed_item_extension and add new extensions
create is__file(), is_image(), is_video() which can tell if one specific extension (from $allowedExtensions) is file or image or video respectively.
get_item_type() uses these 3 previous functions to return the item type with the same name as the folders of the user.
upload_item.php:
modify all the source file. First check if the session index itemToUpload is set and if not prints an error message
else begins the uploading. At first initialize some variables and get the username from $_SESSION index. Checking if the selected item has allowed extension and store the error message to errors array. Then check the size of the item if it's bigger than 20 MB and store the error message in errors array. At the end check this array if it has values if(!empty($errors)) which means errors with the constraints of uploading else call the move_uploaded_file() to move the temporary uploaded item in the correct path "users_items/<username>/<item_type>/<item_name>")" after that puts the successfulUploadedItemMessage in $_SESSION an redirect to home.php.
home.php:
create an message variable which is initiate with empty string after that check if the successfulUploadedItemMessage was set from upload_item.php and if it does gets the value and store it to $successfulUploadedItemMessage message variable. At the end print this variable.