test the undo revert (or cherry pick works) and information update for is__file()
I select the "revert changes in commit" at the last commit an after that do the "cherry-pick commit" if it give the last commmit back. After that push the new changes in upload-items branch and see the result from the web.
Also the is__file() at the etc/functions.php has 2 underscores due to the php built-in is_file() because php cannot override methods.