Add new features in functions.php and upload_item.php
Add
functions.php:
implement the methods get_user_id_by_username() which is get the user id from DB for one specific username, upload_item_in_db() which is insert into the DB the properties of the item and upload_item_in_filesystem() which is upload the item itself in servers filesystem using the function move_uploaded_file().
upload_item.php:
using the get_user_id_by_username() for getting the user id for later use and call the functions upload_item_in_db() and upload_item_in_filesystem() for insert the items properties in DB and item itself in filesystem.
Comment the redirect_to() for not doing this after the upload to see the $query from upload_item_in_db() because it's still under construction.