update functions.php and update_item.php
Add
functions.php: new allowed extension (.txt) for testing with new file the date of upload of the item in DB.
Fix
functions.php: Insert date of upload in DB works. adding semicolons in the $query of upload_item_in_db() the date_of_upload field in DB it is not anymore 0000-00-00 but has the value of current date.