Add content in download_item.php
At first check if the path is set with isset($_GET['path'] and if not prints an error message. Ιn a different case if the variable is defined, the clearstatecache() function is used to clear the cache that was previously stored. The file_exists() function is used to check whether the item exists in the server's filesystem. Next, the header() function is used to set the necessary header information before using the readfile() function. The basename() function is used to retrieve the filename, and the filesize() function is used to read the size of the item in bytes, which will be shown in the download dialog. The flush() function is used to clear the output buffer and after that call the readfile().