Create delete option
Modify home.php: add $deleteItemLink whick has the same declaration structure as $downloadItemLink but the link directs to delete_item.php. After that, shows it at each item as an option.
Add delete_item.php: At first include the functions.php for access some functions later and after check if the path is set with isset($_GET['path'] and if not prints an error message. Ιn a different case if the variable is defined, get the path itself and call the functions delete_item_from_filesystem() and delete_item_from_db() for upload the item itself to server's filesystem and update the database with the new item information. At the end redirects back to home.php.
Update functions.php: create the function delete_item_from_filesystem() that takes the items full path as a parameter and check if this item exists in server's filesystem. If exists delete the item with unlink() function and if an error occurs prints error message else prints a successful message. The function delete_item_from_db() create an SQL query that delete a record in items table with specific item path.