Add style to login.php
login.php: Find the right path to load the login avatar icon.
Change The Welcome to Verobox with Verobox and make it a div instead of h2 tag with welcome-message class.
upload_item.php: Change the message for not allowed items.
style.css:
import a new font from google fonts.
Initiate the body of html with some color the new font an alignment.
Style the login-box div with his size, backgorund color, font color and position in page.
The avatar icon has been resized, circular and moved to the upper center of the login-box.
Move the welcome-message in center and change the font to 29px and bold.
Make the username and password labels bold.
Make all the inputs tags take all the width across the login-box div and add extra margin in bottom.
the inputs that are responsible for the username and password make the background transparent with no outlines except bottom that is solid white with some specific height and font size and color.
The login button (or submit) has no border lines or outlines with stardard size, font size, backround color and font color. Also have rounded edges using border-radius set to 20px.
If you hover the login button change the cursor to pointer, the backround color anf font color of the button.
The a tag (or registration link) discard all the standard decoration and set the fonts to white ,bold and 16px and line-height to 16px.
If you hover the a tag the font color changed.
The class error which is responsible to show error messqages about wrong credentials set the color to red (hex: ff0000).