Make function set_flash_message() and get_flash_message() for showing information messages and minor changes.
Replace the code for handling flash (information) messages via session dierctly with 2 functions. Βecause this code was repetitive at every message i made the function which has the same code but the name of index in seesion is the parameter because only that change every time. In the get_flash_message() define only the index name of the session. In the set_flash_message() define the index name and the information message.
Also replace $_SESSION['uploadedItemStatusMessage'] if it's empty with get_flash_message('uploadedItemStatusMessage') because has the same function at this moment.
Make the flash div (or toast) with specific decoration and the span message inside of it and alert class which put it inside the element with specific id (works but not supported with div element), also with class show and hide using @keyframes move horizontaly with translateX the whole div.
Add in login.php the nesessary javascript for add and remove the classes "show", "alert" and "hide" in the element with the temporary id="myDIV".