logout and registration message toast works
style.css: change the name from toast-logout to toast because now bacause shows and the successful registration message as well. I also changed the colors to softer tones.
home.php, register.php: add a "," to separate the username from the rest message.
login.php: add a div as toast container and inside a span for the message itself which is get it from echo get_flash_message('successfulLogoutMessage'); and echo get_flash_message('successfulRegistrationMessage'); which only one of the two will have message. Next, link the external javascript file for calling functions. For show and hide the toast call the function toast() with "toastId" as parameter which run on onload event with specific element id. This event doesn't support in div tag but it support in body tag. It's the same because one way or another will run at the load of the page.
functions.js: add the toast function which show or hide the div with elementId. First get the text from the span tag which belongs to message class. If doesn't has message means no logout or registration occurs show in console an error message else run the commands for show/hide the toast. First, get the div (with elementId) and add the classes show and alert for showing the toast (slide left with css) and remove the class hide and after 3 seconds remove the class show and add the class hide for hiding the toast (slide write with css).