styling the username and password error messages
login.php: Change the error messages and their span to div. Remove the labels for username and password.
style.css: Change the height of login-box. Change the thickness of left border from 4 to 5 pixels and give circular shape to the edges. I modified the whole error class which i give all her elements a specific position to the login-box, some background color, font color and text alignment. Also i round the top edges of the elements. The elements of this class looks like one error area with error messages inside.