Finish styling login page
Add a favicon.ico in root folder for the browser image. Μost browsers add it automatically when its in root folder. No need for extra code.
login.php: In head i add the link for font awesome icons. For the username input add a user icon and for the password input add a key icon from font awesome. I wrapped the combination of icon and input with a div element. Also i keep only one error div (rename to input-error) outside of input-box which print the username error and password error together. But only 1 of the 2 can have a value at the same time and the other is empty.
style.css: Change the edges of login-box for to be less round. Add command for the user not to select the text from welcome message and error message. Rename the class message to toast-message. Add shadow underneath the toast. Add some margin under the submit button and the width to 100% of the parent.
Add to the class fa from font awesome some padding from left as a recess.
Remove the white line from input to class input-box to underline both icon and text and add some spacing to the elememts. Change the width of the inputs to 85% and add some left padding to have a distance from the icon.