Add styling register.php with css
Add register avatar (PNG image) for register.php

Add in register.php: Move the head element after the top php code and add to it some meta elements and the link for font awesome icons and the javascript functions script. Also add the register-box which is used as a container for all the components. Add the avatar path and the register-message class. Inside of the form add 2 div elements one for the left side and one for right side which are used as containers for the elements to be placed on the left side and the ride side. The left container conains the elements for first name, last name, username and password and the right container contains the elements for email date of birth and sex of the user. At the end is the register button. The input-error and link for the login page is outside of the form.

Add in style.css: import some fonts from google, add the body element, classes for register-box, register-avatar, register-message, input-box, input[type="text"], .register-box input[type="password"], input[type="submit"], input[type="submit"]:hover, ::placeholder and input-error which is the same as login one's with some minor changes to pair with the left and right components.

Extras: delete unused background for login.php