style the left and right containers
Modify: the position of and size of left div and to the left of the border one line as a divider between the 2 div elements. Also modify the size and position of the right div. Change the size and posistion of the register button and the same properties for login link. The class required-fields reffers to asterisk and  * Field is required label. The class required-fields-label reffers only to * Field is required label.