Make search bar
functions.php Move the code which is responsible for display the data to screen from home.php to functions.php as a function named "display_items". Also rename the function "get_all_user_items" to "get_user_items". I modified this function because it has the same logic as a search function making only one change to the query with one extra parameter which if the value of this parameter is "" (empty string) the function get all the user items (due to % characters in MySQL) and if the string has some value the function returns all the items who has this specific value in any position in item_name (have the ability to add more fileds of the table later).
home.php: add the get_toast_message() which searchItemStatusMessage index is responsible on whether the user click the search button leaving the text input blank or not. Also add the search form and name the name attribute of submit button to "submit-search" and for this reason rename the name attribute of submit button of upload to  "submit-upload". initialize the parameter  $itemToSearch to "" (is the extra parameter of searching) and check if the POST index itemToSearch is set (or more simply the user click the search button). If not set the variable $itemToSearch remains an empty string and the get_user_items() gets all the user items. If the index itemToSearch is set check for empty string and if it is call the function set_toast_message() passing the string error message "Type something to search". If the index itemToSearch is not empty change the value of the variable $itemToSearch  with the value from search input text and the function get_user_items() will return all the user items who have the value of the variable $itemToSearch in any position in their name. At the end call the functions display_items() to display the results from get_user_items().