Add search options (after POST of search the search option dont work)
home.php: Add checkboxes in search form which act as extra search options. Remove the unessesary   if(isset($_POST['itemToSearch'])) because the empty already check if is set the index. Also add an if check for possible checked checkboxes (only extension checkbox for now) and get the extension from search input text and with this extension as a parameter call the function filtering_items_by_extension() to modify the current $items (which is all the user items at the moment).
functions.php: add the png extension in allowed extensions for upload (also in function is_image()). Add the function is_checked_checkbox() to validate if one checkbox is checked.
Add the function filtering_items_by_extension() to hold only the items with the specific $extension.
Also change the code of function get_item_extension().