Fixed: (after POST of search the search option dont work)
The problem was how the searchOption[] parameter passed in function is_checked_checkbox. No need for [].
Modify functions.php: Change the function get_item_extension() to the previous code because caused a notice error for the parameter.
Also change the code of function is_checked_checkbox() to a more simpler code.
At the end change the name of function filtering_items_by_extension() to filter_items_by_extension().
home.php: Add the necessary code for the checkbox to remain checked after searching with the function is_checked_checkbox() (if was selected).