home.php: fixed the search error issue
Fixed: change the if statement with a more accurate one (includes the search button itself if clicked, not only the empty input text). The first if statement checks if the search button pressed and the input text is empty. If that the case set a toast search error message and reload the page immediately.
If this not the case do a second check which this time check if the search button was pressed and the input text wasn't empty. In this case get the value from the input text and continue what you were doing.
Extra: Remain the text in search's input text after the search.