Add the video search option
The steps for adding in search options the option to show only videos is the same as images and files. Add an input radio in search form with the nesessary code to remain check from previous selection and an if statemen to check again if the radio is selected and call the function filter_items_by_type() to filter the items and keep those who belong to video item type.