Make search work correctly with checkboxes
Make the input radio input checkbox again.
Initilize an array named $items which contain the items to be displayed to the screen.
Now the user items stored in the array $allUserItems which continues to have the items with the keyword from searchbar in any position in the item name. This array acts as a reference array which the additional search options get the values from. In this way each additional option filters items from all user items and not from already filtered items due to some previous selected search option.
The steps performed for this process are:
     If the extension checkbox was selected
          -get the items extension
          -filter from all the user items ($allUserItems) the items that they have the specific extension
           -Add (append) the result in the $items array with the function array_merge(). Now this array contains all the items with this extension.
        -
In this way in the array $items are appended the results for every search option at the same time instead of one search option each time.
     If the file checkbox was selected
          -filter from all items ($allUserItems) the items that they belong to file type.
           -Add (append) the result in the $items array with the function array_merge().
Now the array $items contains all the previous values (from extension checkbox) and current values.
The same principle applies to the following checkboxes if they are selected.
Because the array that displayed is the array $items and if you load the home.php without search something the the $items has no value. In this case make an if statement where it checks if the search button was clicked. If not clicked put the value from the $allUserItems to the $items.