Fixed: If not select any search option doesn't display any item (like you haven't any)
Before if didn't select any search option and click "Search" the toast error message was displayed but no items were displayed like you haven't any uploaded item. That because of if statement if(!isset($_POST['submit-search'])) where the approach was wrong. The corrent statement was to pick the correct POST index where was the "searchOptions". I change the structure to embody the if statemants which are responsible for the additional search options. The final result is if the index searchOptions is set (select atleast one search option checkbox) continue with the corresponding tasks. Else remain with the values of variable $allUserItems which act one keyword (value from input text) search.