Style home page: Add the verobox logo, Modify the search bar and add a logout icon
home.php:
Style (class: home-logo) the verobox logo which act as a loading home page button. Style the logout link which has the form of a power off button. Belongs in the class logout and contains one power-off icon from font-awesome.
style.css: add height to menu bar (50px).
Style the class home-logo to be in the left side (5.5% of the menu). Also style the tag a inside of it which has no text decoration, specific font size and color.
Style the whole search bar (include the search options) to be at the center of the menu (same method as the home-logo but in center, 50% of menu). Style the logout class tο be at the right of the menu (same method as the home-logo but in right, 97% of menu). Also style the font-awesome icon to have white color and specific size.