Add home-content area and sidebar
Undereath menu add 2 div's.
At the left side add the side bar and at the right side the home (main) content area to display the items.
home.php:
add the sidebar div with class "side-bar".
Add the home content area with class "home-content" and throw inside the upload form, all the php code for additional search options and display items section.
style.css:
add the body tag and except margin and padding equals 0 add the property overflow: hidden to not show any scroll bar for any resizing of the browser window.
The class "home-content" has no outline and border to fit excaclty at the edges of window. Also has no float and fixed position and with left, width, top, height properties place the div at the right from the bottom of menu till the end of body tag with white (#fff) color. The property "overflow-y: auto;" means the that show up a scroll bar if the content exceeds the size of window.
The class "side-bar" has neither outline, border or float but has the fixed position, the same height and top position with the "home-content" and different left position and width so as to be placed in the gap next to "home-content" with black color.