Style search-bar with drop down search options
home.php: Change the input submit tag to button submit tag to add as text the search icon from font awesome because the button tag has close tag and input hasn't.
style.css: make verobox logo bold.
Make the form a single unit that contains search options drop down menu at the left the input text for the user inputs and the button with the search icon at the right.
Make the form size with display flex.
Make the size of search_options_dropdown class (with cursor pointer).
Also make the before property to make an expand arrow using border.
Make the class selected_search_options for the text of the search dropdown list.
The ul class make the list of the items (which is search option checkboxes) (size and positioning under the search_options_dropdown)
The class show-list (not used) is for later to show and hide the list with toggle javascript function
the css for li tag is to have space under each list item (checkbox)
The property last-child is for the last item to undo the spacing for beter looking of list (space at the end without meaning) make the search_field class with specific size
Also make the input text styling, size and position and its placeholder color.
Also make the submit button styling, size and position and the hover event to change background color.