Make the dropdown menu working when clicking on "search options"
js/functions.js: make the function show_search_options_dropdown_list() which gets the element with the specific class name and toggle the class "show-list" ("toggle" means that when you click the "selected_search_options" it adds the class "show-list" to the "dropdown-list" class and if you click again it removes it).
home.php: Add the click event on div with class "selected_search_options" and id this event occurs execute the function show_search_options_dropdown_list() passing the parameter "dropdown-list" which corresponds to that element with the specific class means this function modify this element.
style.css: Changed the ul element with its class (dropdown-list). Also remove the ul from ul .show-list because the class show-list is at the same level with the ul element. At the end increase the width of dropdown list to 135px to match the div above it.