Style the search bar
style.css:
At the input text make the font bold with 1px space between letters but keep the placeholder the way it was (normal).
Also to make the "dropdown-list" appear as an integral part of the "search-bar" i remove the border-radius from the top edges of the list and place it slightly to the top (top :38px -> top: 33px).
home.php:
add some spaces in the code between MENU BAR, SIDE BAR and HOME CONTENT.