Style search-bar (options list)
Transform the default checkbox  to look like an on/off switch.
First, i remove all the default appearence and resize the element. Add shadow and transition delay if it comes back to this form. If the user check the checkbox change the color to DodgerBlue.
With ::before pseudo element create an new content which has a circular form. Add some color, shadow and transition time to go smoothly from one side to the other.
At the end with :checked and ::before which means where to stop after select