Move the upload form in the sidebar
home.php: In the sidebar add a list for menu (using "ul" and "li", li class = "list-item"). As first list item add the item upload form.
style.css: change the height of menu (NOT FIXED the search items dropdown list). Also change the backgorund color of the sidebar and for the list-item class add the "list-style: none;" which removes all the styles of the list.