Add upload item progress bar
style.css: Remove all the margin and padding from each sidebar element. Made the form block display so that the upload button to be bellow the item selector. Make the class "upload-item-selector" a white box to display the selected item to be upload. With the pseudoclass "-webkit-file-upload-button" transform the "choose" button to be a yellow-ish larger button and the cursor from above has the form of pointer. Also make the "Upload" button (class "upload-btn") a blue-ish button to takes up the whole width of the sidebar. At the end tranform the div who is responsible for the progress bar with specific height 100% width of the sidebar and a blue-ish loading color.
home.php: Named the upload items form with the class name "upload-items-form" and the tag input file with the class nane "upload-items-selector". Assosiate the upload button with the class "upload-btn". Add the div with class name "progress-bar" which by default has display=none and a transition width time. The comment (with script tag) contains the necessary code to upload with javascript the item and get the required data to work the progress bar (the progress bar is loading but didn't upload the item).