Add the files upload box
home.php:
sidebar: Remove the upload form from the side bar and add a single button "+NEW" to open the upload-box (class="open-upload-box-btn").
home-content: add the div with class name "upload-box" whick contains a header message (1) the upload form (2), a progress area that shows the currently uploading item with some information of the item and an uploading progress bar (3) and an uploaded area which contain some information of the item too (4).
Number explanation:
1) The header tag contains the title of upload-box.
2)The form contains and font-awesome icon and a p tag "Browse File to Upload" to explain to the user that this is the form to click and select item.
3) The progress area is a box which contains all the files that currently uploading. Has a font-awesome file icon and the content which contains the item name and the percentage of the loaded
item at this time and a progress bar with this value.
4)The uploaded area is the same thing as the progress area but instead of the percentage and the progress bar has the font awesome check icon and the item size.
style.css:
In the class list-item of the side bar add the flex display, centered the elements and add a border at the bottom which acts as a divider.
Commented all the classes which are related with the old sidebar upload form.
The class .side-bar .list-item .open-upload-box-btn is for the new open button to make it circular with a shadow and a pointer cursor. Also have a space-evenly justify-content and flex display to show the content of the button straight.
The class  for the upload-box make the element a rectagle form with some padding which contain the require information and place it in the middle of the screen (left = 42% instead of 50% because is the upload-box is located inside the home-content, a little further left).
Make the header a little bigger and with some color.
The class ".home-content .upload-box form" make the form a box with dashed border and centered the font awesome cloud icon and the text.
Inside the form color the element where the element is i or p tag. Then for each item change the font size.
Make the row of uploading box a light purple box with flex display and center align items.
The .home-content .upload-box section .row i is responsible for the font awesome file icon for decoration and the details span section for font size.
The progress area content has width 100% to the parent and a margin left for spacing from the font awesome icon.
the details has flex display and a space between the content.
The class progress-bar has 100% width and specific height with white color and border radius.
The class progress is the filling part with 100% height of the progress-bar different color, inherit border radius (same as parent element) and the width which will change with the javascript and will have the effect of loading.
The uploaded area contains the uploaded history for this reason has max-height of 150px for possible overflow (this height shows without scroll about 2 uploaded items)
Make the scroll bar invisible without affect the row's width using width 0.
The content has flex display and center align items.
THe details has flex display and column flex-direction which place the items the one below the other and add some spacing with margin left
The class size has specific font size and color.
Change the font awesome check icon to a specific font size.