Add: open upload dialog after click the +new button from sidebar
style.css: make display property none so that it is constantly hidden.
home.php: I added to the sidebar's upload button (+new) the the handler for onclick event where calls the javascript function show_upload_dialog() with the className parameter which is the class name "upload-box".
functions.js: Add the function show_upload_dialog(className) which gets the element with the specific className and change from the style the property display to "block".