Add: items are uploaded using ajax
Passing the items to upload_item.php using AJAX.

style.css: add the cursor property to pointer inside the upload form and the user select: none to not select the text "Browse File to Upload". Also remove the commented code from the previous design.
home.php: add a hidden input file tag with class name "upload-item-selector". Remove the existed content in the progress-area and uploaded-area. Change the "file" words with "item" words in texts.
upload_item.php: comment all the extra php code (filters, toast messages etc.) to run only the basic code for uploading a file to be compatible with ajax.
functions.js: In the function show_upload_dialog() initiate all the necessary variables and objects to have a complete oparating upload box from the moment the +new button is pressed. Get the form from upload-box as an object and initiate some variables for later processing. Also add an event listener on click event in form which get the upload item selector (input file tag) and get the "progress-area" and "uploaded-area" to add the uploading item and the history of uploaded items and then perform a click action in JavaScript to open the item selector. After that i add one more event listener onchange and take the selected item (the 1st if multiple items selected). If an item selected and was not canceled take the item name and split it in some form if it is longer than 30 characters. At last call the function upload_item and pass the necessary arguments to be used from by her. (itemName, uploadForm, progressArea, uploadedArea). When function upload_item executed create a new XMLHttpRequest object (AJAX) and open a connection with post method to upload_item.php. In upload i add an event listener (progress) while getting the loaded and total variables. This event handler gets the itemPercentageLoaded and itemTotal (KB). The item size is a variable to calculatd if the item's size will appear in KB or MB. Because the itemTotal is already in KB if it's less than 1024 (KB) we add the string " KB" to specify the measurement unit. Otherwise devide the variable loaded (1024 * 1024, to convert to MB) with 2 decimal points. After that i built the progressHTML which contains the HTML script containing all elements and classes to show the uploading information div (progress-area) but replace with JavaScript the itemName and itemPercentageLoaded to show the selected item. In div "progress" add the style property directly which change the width of this div to act as a progress bar and add the progressHTML with innerHTML to the progress-area section. If the item has been loaded (loaded == total) then set the progressArea to a null string to show nothing and initiate the uploadedHTML which contains all the elements and calsses to show the uploaded item in a div but with JavaScript replace the itemName and itemSize. After that with line "uploadedArea.insertAdjacentHTML("afterbegin", uploadedHTML);" add the complete uploaded items in a way to look like a list (one on top of the other). At last create a FormObject with the parmeter of uploadForm to easily send forms with AJAX and with function send, the data are transferred to upload_item.php.