Recreate some php functions to JavaScript functions
This functions has the role of filters and did this to execute these filters in client side. Equivalent filters in php are executed in server side (after the upload of the item which is nonsense).
These filters are:
-get_item_extension()
-is_allowed_item_extension()
-is_file()
-is_image()
-is_video()
-get_item_type() (didn't used it so far)

new max_item_size_allowed()

Also attempt to execute these filters in the same way as I did in upload_item.php before the POST of the data.