Add: Display "byte" item size
I previously manage the items directly as "kilobytes". Now i handle them as "bytes". This information is held by the variable "total". The variable itemSizeInBytes is the renamed variable total.
If the value of the variable itemSizeInBytes is less than 1024 remain in B (bytes) else it is from "kilobyte/KB" and above. Then if it's less than Math.pow(1024, 2) (1024 * 1024 = 1.048.576 bytes or Megabyte) it's KB. Else its in MB.