Add: close button to upload-box
home.php: add a button tag which belongs to class "close-upload-box-btn" and trigger the javascript function hide_upload_dialog() with "onclick" event and passing the 'upload-box' class as a parameter. The button contains the "fa-times" font-awesome icon to appear as an exit icon.
functions.js: get the element from the className parameter which is the "upload-box" and set the css display to "none" to make it disappear.
style.css: with hover event set the background-color to #f4f4f4 to show the button is highlighted. Also make the class "close-upload-box-btn" which has position absolute, float none and top, left and tranform fields the desired values so that the close button appears on the upper right of upload box. Also make the button circular, added some color and make the cursor pointer.