add function to reload a specific html element
home.php: the div with class "home-content" gets an id with name "home-page".
functions.js: add the function update_element() which gets the element with "home-page" id and reload it using jquery. I also nested the progressHTML with the corresponding if stamement (loaded < total) to run only in correct cases.