Add a box display for items
home.php: Add a div "grid-container"which contains the php function display_items().

functions.php: In the function display_items() add an if statement for the item name where if the name length is greater than 35 characters long then shorten it using the first 31 characters and the extension of the item with the function get_item_extension(). After that using the "echo" keyword display a div "grid-item" which contain the div "item-header" and the div "item-information" which also contain the div "item-name".

style.css: add the class grid -container which have a grid display with 4 columns which one takes the 22% of the entire space of the grid with space-evenly justify-content. Also have 30px grid gap between items a white background-color and 19px padding for the grid. The class "grid-item" contains the properties of every item (header and information section) have realtive position 2px solid border with ligthgrey color specific height and some border-radius. The "item-header" temporary some text but will contain a thumbnail or font-awesome icon depending on the item type. This takes the 70% of the height of the grid-item from the top and have some border-radius to match the top corners of the grid-item. The "item-information" contains the name of the item, takes the bottom 30% of the grid-item height and have border-radius to match the bottom corners of the grid-item. The div "item-name" place the text in the left side with some left space in the middle in Y axis with specific font size and weight.