Update the look of the grid-item
functions.php: add the div "icon" which contain the font-awesome icon depending the item type. The item type (based on font-awesome icons name) is included in the icon name.
style.css: the class icon make the font-awesome icon to align in center in Y axis. Edit the "fas" class from the font-awesome icon to specify the color and the size of the icon.