Add div to display the item size
functions.php: add the div "item-size" to display the size of the item derived from the database (there is no such a field in DB at this moment). Also change the length of the item-name again.
style.css: with the class item-size adjust the position of the element just below the item-name and change the font-size. To look nice adjust the position and size of the item-name too (larger font and slightly above).