Add the item size field to the database and accessed it
upload_item.php: get the size (in bytes) of selected item from the global variable $FILES and add it as new parameter to the function upload_item_in_db().
functions.php: In the function upload_item_in_db() get the new parameter $itemSize and add it to the $query variable. Specifically the item_size located in INSERT INTO  is the name of the field in database and in the VALUES  section of the $query is the parameter itself. So far with this process I add the item size to the database. To get the item size from database in the function get_user_items() all i do is include the item_size field of the database. The returned value pass as parameter to the function display_items(). To display the item size in the grid-item all i have to do is for every $item of the parameter $items to choose the item_size and this value to put it on the div "item-size".