Convert the item size unit
upload_item.php: execute the function corvert_item_size_unit() which takes the item size as argument (number) and return the item size after convert the unit (string) (Byte, KByte, MByte etc.).
functions.php: add the function corvert_item_size_unit(). Initiate the $convertedSize with some error value (items with negative bytes don't exist). In first if statement check if $sizeInBytes is less than 1024 bytes and if its true make the $sizeInBytes  string using strval() built-in function and append the corresponding " B" for bytes. Else (1024 and above) if its less than 1048576 bytes (or MB) devided with 1024 to make it KBytes, round it with two decimal number and then make it string and append the corresponding " KB" for the KBytes. For the MB and GB the same principle is followed but with the corresponding divisors and sizes.