Display the item's date of upload and fix the displayed shortened item name
functions.php: The if statement for the $itemName length it's not include the whole basename anymore but only the name (without the extension). Also rename the class "item-size" to "item-date-and-size" and include the $itemDateOfUpload in the same area with the item size.
style.css: Rename the class "item-size" to "item-date-and-size".